pyproject.toml
[project]
name = "agentic-hackathon"
version = "0.1.0"
description = "Agentic hackathon"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"browser-use",
	"python-dotenv",
]
